+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Shellout now requires Gleam v0.26 or later.
+- Shellout now supports the Deno JavaScript runtime (v1.30 or later).
+
 ## v1.0.0 - 2022-06-01
 
 - Complete rewrite!